requirements.txt
asgiref==3.2.10
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
cffi==1.14.4
channels==3.0.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
olefile==0.46
Pillow==7.2.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
PyPDF2==1.26.0
pytz==2020.1
qrcode==6.1
reportlab==3.5.53
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.2.0
